[Implementation] Added Authentication

+++ TaniumApi/Controllers/CommunityController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using TaniumApi.Authentication.Interfaces;
 using TaniumApi.Library.DataAccess.Interfaces;
 using TaniumApi.Library.Models;
 using TaniumApi.Models;
@@ -6,9 +8,14 @@ using TaniumApi.Models;
 namespace TaniumApi.Controllers;
 [Route("api/[controller]")]
 [ApiController]
-public class CommunityController(ICommunityData communityData, ILogger<CommunityController> logger) : ControllerBase
+[Authorize]
+public class CommunityController(
+    ICommunityData communityData,
+    IAuthService authService,
+    ILogger<CommunityController> logger) : ControllerBase
 {
     private readonly ICommunityData _communityData = communityData;
+    private readonly IAuthService _authService = authService;
     private readonly ILogger<CommunityController> _logger = logger;
 
     [HttpGet]
@@ -16,7 +23,7 @@ public class CommunityController(ICommunityData communityData, ILogger<Community
     {
 		try
 		{
-			var communities = await _communityData.GetAllCommunitiesAsync();
+            var communities = await _communityData.GetAllCommunitiesAsync();
 
 			return Ok(communities);
 		}
@@ -28,6 +35,7 @@ public class CommunityController(ICommunityData communityData, ILogger<Community
     }
 
 	[HttpGet("{id}")]
+	[AllowAnonymous]
 	public async Task<IActionResult> GetCommunityAsync(int id)
 	{
 		try
@@ -48,6 +56,7 @@ public class CommunityController(ICommunityData communityData, ILogger<Community
 	}
 
 	[HttpPost]
+	[AllowAnonymous]
 	public async Task<IActionResult> CreateCommunityAsync(CreateCommunityModel body)
 	{
 		try
@@ -57,14 +66,20 @@ public class CommunityController(ICommunityData communityData, ILogger<Community
 				return BadRequest(ModelState);
 			}
 
-			// TODO: Add user auth
+            var loggedInUser = await _authService.GetUserFromAuthAsync(HttpContext);
+            if (loggedInUser is null)
+            {
+                return StatusCode(401, "Unauthorized");
+            }
 
-			var community = new CommunityModel()
+            var data = new CommunityModel()
 			{
-
+				UserId = loggedInUser.Id,
+				Name = body.Name,
+				Description = body.Description,
 			};
 
-			var createdCommunity = await _communityData.CreateCommunityAsync(community);
+			var createdCommunity = await _communityData.CreateCommunityAsync(data);
 
 			return Ok(createdCommunity);
 		}
@@ -85,14 +100,33 @@ public class CommunityController(ICommunityData communityData, ILogger<Community
                 return BadRequest(ModelState);
             }
 
-			var community = new CommunityModel()
+            var loggedInUser = await _authService.GetUserFromAuthAsync(HttpContext);
+            if (loggedInUser is null)
+            {
+                return StatusCode(401, "Unauthorized");
+            }
+
+			var fetchedCommunity = await _communityData.GetCommunityAsync(body.Id);
+			if (fetchedCommunity is null)
+			{
+				return NotFound();
+			}
+
+
+			if (fetchedCommunity.UserId != loggedInUser.Id)
+			{
+				return StatusCode(401, "Unauthorized");
+			}
+
+            var data = new CommunityModel()
 			{
 				Id = body.Id,
 				Name = body.Name,
 				Description = body.Description,
+				UserId = loggedInUser.Id,
 			};
 
-            var createdCommunity = await _communityData.CreateCommunityAsync(community);
+            var createdCommunity = await _communityData.CreateCommunityAsync(data);
 
 			return Ok(createdCommunity);
         }
@@ -108,6 +142,23 @@ public class CommunityController(ICommunityData communityData, ILogger<Community
 	{
 		try
 		{
+            var loggedInUser = await _authService.GetUserFromAuthAsync(HttpContext);
+            if (loggedInUser is null)
+            {
+                return StatusCode(401, "Unauthorized");
+            }
+
+            var community = await _communityData.GetCommunityAsync(id);
+			if (community is null)
+			{
+				return NotFound();
+			}
+
+			if (community.UserId != loggedInUser.Id)
+			{
+                return StatusCode(401, "Unauthorized");
+            }
+
 			await _communityData.DeleteCommunityAsync(id);
 			return Ok("Success!");
 		}
